+++ GenericModConfigMenu/Framework/Overlays/KeybindOverlay.cs
@@ -1,56 +1,87 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using GenericModConfigMenu.Framework.ModOption;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using SpaceShared.UI;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
 using StardewValley;
+using StardewValley.Extensions;
 using StardewValley.Menus;
 
 namespace GenericModConfigMenu.Framework.Overlays
 {
-    /// <summary>Manages an active keybind overlay.</summary>
-    internal class KeybindOverlay<TKeybind> : IKeybindOverlay
+    /// <summary>An overlay which lets the player view and edit a keybind list.</summary>
+    internal class KeybindOverlay
     {
         /*********
         ** Fields
         *********/
-        /// <summary>The keybind option being bound.</summary>
-        private readonly SimpleModOption<TKeybind> Option;
+        /****
+        ** Constants
+        ****/
+        /// <summary>The width of the overlay box, including the <see cref="ContentPadding"/>.</summary>
+        private const int BoxWith = 650;
 
-        /// <summary>The label to update with the bound keys when <see cref="Option"/> is rebound.</summary>
-        private readonly Label Label;
+        /// <summary>The pixel spacing around the content within the box.</summary>
+        private const int ContentPadding = 30;
+
+        /// <summary>The left pixel indent of the keybind list, added to the <see cref="ContentPadding"/>.</summary>
+        private const int KeybindListIndent = 40;
+
+        /// <summary>The button action to save changes.</summary>
+        private const string OkAction = "OK";
+
+        /// <summary>The button action to clear the keybind list.</summary>
+        private const string ClearAction = "Clear";
+
+        /// <summary>The button action to add a keybind to the list.</summary>
+        private const string AddAction = "Add";
+
+        /// <summary>The button action to remove a keybind from the list.</summary>
+        private const string RemoveAction = "Remove";
+
+        /****
+        ** State
+        ****/
+        /// <summary>The keybinds to edit and save.</summary>
+        private readonly List<Keybind> Keybinds;
+
+        /// <summary>Whether the config only allows a single button, rather than a full keybind list.</summary>
+        private readonly bool OnlyAllowSingleButton;
+
+        /// <summary>The translated keybind name.</summary>
+        private readonly string Name;
+
+        /// <summary>The callback to invoke with the updated keybinds when the overlay is closed and saved.</summary>
+        private readonly Action<Keybind[]> OnSaved;
+
+        /// <summary>The display text to show within the content.</summary>
+        private readonly List<ClickableComponent> Labels = [];
+
+        /// <summary>The clickable buttons in the form.</summary>
+        private readonly List<ClickableTextureComponent> Buttons = [];
 
         /// <summary>The pixel position and size of the keybind UI.</summary>
         private Rectangle Bounds;
 
-        /// <summary>The button which keeps the current keybind.</summary>
-        private ClickableTextureComponent OkButton;
-
-        /// <summary>The button which clears the keybind when clicked.</summary>
-        private ClickableTextureComponent ClearButton;
-
         /// <summary>Whether to reset the layout on the next update tick.</summary>
         /// <remarks>This defers resetting the layout until the menu is drawn, so the positions take into account UI scaling.</remarks>
-        private bool ShouldResetLayout;
-        
-        /// <summary>Whether a button has been pressed since the menu has opened.</summary>
-        /// <remarks>This prevents checking for keybinds before the menu has opened, since the menu is opened with a key press.</remarks>
-        private bool HasPressedButton;
+        private bool ShouldResetLayout = true;
 
-        /// <summary>List of buttons currently held.</summary>
-        /// <remarks>This keeps the held buttons in the order they were pressed.</remarks>
-        private List<SButton> PressedButtons;
+        /// <summary>Whether the displayed keybind changed.</summary>
+        /// <remarks>This is equivalent to <see cref="ShouldResetLayout"/>, but doesn't prevent the user from clicking UI elements (e.g. so clicking 'OK' doesn't start a keybind registration for <c>MouseLeft</c>).</remarks>
+        private bool ButtonsChanged;
+
+        /// <summary>The keybind actively being edited by the player, if any.</summary>
+        private KeybindEdit KeybindEdit;
 
 
         /*********
         ** Accessors
         *********/
-        /// <inheritdoc />
+        /// <summary>Whether the player has finished binding the key, so the overlay can be closed.</summary>
         public bool IsFinished { get; private set; }
 
 
@@ -58,113 +89,95 @@ namespace GenericModConfigMenu.Framework.Overlays
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
-        /// <param name="option">The keybind option being bound.</param>
-        /// <param name="label">The label to update with the bound keys when <see cref="Option"/> is rebound.</param>
-        public KeybindOverlay(SimpleModOption<TKeybind> option, Label label)
+        /// <param name="keybinds">The keybinds to edit and save.</param>
+        /// <param name="onlyAllowSingleButton">Whether the config only allows a single button, rather than a full keybind list.</param>
+        /// <param name="name">The translated keybind name.</param>
+        /// <param name="onSaved">The callback to invoke with the updated keybinds when the overlay is closed and saved.</param>
+        public KeybindOverlay(Keybind[] keybinds, bool onlyAllowSingleButton, string name, Action<Keybind[]> onSaved)
         {
-            this.Option = option;
-            this.Label = label;
+            this.Keybinds = [.. keybinds];
+            this.OnlyAllowSingleButton = onlyAllowSingleButton;
+            this.Name = name;
+            this.OnSaved = onSaved;
 
-            this.ShouldResetLayout = true;
-            this.HasPressedButton = false;
-            PressedButtons = new();
+            // go straight to edit view if needed
+            if (onlyAllowSingleButton || keybinds.Length == 0)
+                this.StartEditingKeybind(0);
         }
 
-        /// <inheritdoc />
+        /// <summary>Try to handle the pressed buttons, either by assigning the keybind or cancelling the UI.</summary>
+        /// <param name="e">The event arguments.</param>
         public void OnButtonsChanged(ButtonsChangedEventArgs e)
         {
-
-            // add pressed keys
-            if (e.Pressed.Any())
+            // exit menu
+            if (e.Released.Contains(SButton.Escape))
             {
-                HasPressedButton = true;
-                PressedButtons.AddRange(e.Pressed.Where(b => this.IsValidKey(b) && !PressedButtons.Contains(b)));
+                Game1.playSound("bigDeSelect");
+                this.Exit(save: false);
+                return;
             }
 
-            // check if a key has been pressed since opening the menu
-            if (!HasPressedButton)
-                return;
-
-            // get released keys
-            SButton[] released = e.Released.Where(this.IsValidKey).ToArray();
-
-            // apply keybind
-            if (released.Any())
+            // handle keybind edit
+            if (this.KeybindEdit != null)
             {
-                this.HandleButtons(PressedButtons.ToArray());
-                this.IsFinished = true;
+                // add new keys
+                if (this.KeybindEdit.Add(e.Pressed))
+                    this.ButtonsChanged = true;
+
+                // finish on key release
+                if (this.KeybindEdit.Any() && e.Released.Any(this.KeybindEdit.IsValidKey))
+                    this.FinishEditingKeybind();
             }
         }
 
-        /// <inheritdoc />
+        /// <summary>Update the overlay when the window is resized.</summary>
         public void OnWindowResized()
         {
             this.ShouldResetLayout = true;
         }
 
-        /// <inheritdoc />
+        /// <summary>Handle the player left-clicking the overlay.</summary>
+        /// <param name="x">The pixel X position where the player clicked.</param>
+        /// <param name="y">The pixel Y position where the player clicked.</param>
         public void OnLeftClick(int x, int y)
         {
             if (this.ShouldResetLayout)
                 return;
 
-            if (this.ClearButton.containsPoint(x, y))
+            // clicked button
+            foreach (ClickableTextureComponent button in this.Buttons)
             {
-                Game1.playSound("coin");
-                this.SetValue(Array.Empty<SButton>());
-                this.IsFinished = true;
+                if (button.containsPoint(x, y))
+                {
+                    this.PerformButtonAction(button.name);
+                    return;
+                }
             }
-            else if (this.OkButton.containsPoint(x, y) || !this.Bounds.Contains(x, y))
-            {
-                Game1.playSound("bigDeSelect");
-                this.IsFinished = true;
-            }
+
+            // clicked out of bounds
+            if (!this.Bounds.Contains(x, y))
+                this.PerformButtonAction(KeybindOverlay.OkAction);
         }
 
-        /// <inheritdoc />
+        /// <summary>Draw the keybind overlay.</summary>
+        /// <param name="spriteBatch">The sprite batch being drawn.</param>
         public void Draw(SpriteBatch spriteBatch)
         {
             // reset layout if needed
-            if (this.ShouldResetLayout)
-            {
+            if (this.ShouldResetLayout || this.ButtonsChanged)
                 this.ResetLayout();
-                this.ShouldResetLayout = false;
-            }
 
             // background
             spriteBatch.Draw(Game1.staminaRect, new Rectangle(0, 0, Game1.uiViewport.Width, Game1.uiViewport.Height), new Color(0, 0, 0, 192));
             IClickableMenu.drawTextureBox(spriteBatch, this.Bounds.X, this.Bounds.Y, this.Bounds.Width, this.Bounds.Height, Color.White);
 
-            // "Rebinding key:" text
-            {
-                string str = I18n.Config_RebindKey_Title(this.Option.Name());
-                int strWidth = (int)Game1.dialogueFont.MeasureString(str).X;
-                spriteBatch.DrawString(Game1.dialogueFont, str, new Vector2(this.Bounds.Center.X - (strWidth / 2), this.Bounds.Y + 20), Game1.textColor);
-            }
-
-            // instruction text
-            {
-                string str;
-                if (!PressedButtons.Any())
-                {
-                    str = typeof(TKeybind) == typeof(KeybindList)
-                        ? I18n.Config_RebindKey_ComboInstructions()
-                        : I18n.Config_RebindKey_SimpleInstructions();
-                }
-                else
-                {
-                    str = typeof(TKeybind) == typeof(KeybindList)
-                        ? string.Join(" + ", PressedButtons)
-                        : PressedButtons.Last() + "";
-                }
-                int strWidth = (int)Game1.dialogueFont.MeasureString(str).X;
-                spriteBatch.DrawString(Game1.dialogueFont, str, new Vector2(this.Bounds.Center.X - (strWidth / 2), this.Bounds.Y + 100), Game1.textColor);
-            }
-
+            // text labels
+            foreach (ClickableComponent label in this.Labels)
+                spriteBatch.DrawString(Game1.dialogueFont, label.label, new Vector2(label.bounds.X, label.bounds.Y), Game1.textColor);
 
             // buttons
-            this.OkButton.draw(spriteBatch);
-            this.ClearButton.draw(spriteBatch);
+            foreach (ClickableTextureComponent button in this.Buttons)
+                button.draw(spriteBatch);
         }
 
 
@@ -174,75 +187,230 @@ namespace GenericModConfigMenu.Framework.Overlays
         /// <summary>Reset the UI layout and positioning.</summary>
         private void ResetLayout()
         {
-            Vector2 pos = Utility.getTopLeftPositionForCenteringOnScreen(width: 650, height: 200);
-            int x = (int)pos.X;
-            int y = (int)pos.Y;
-            int width = 650;
-            int height = 200;
+            // reset
+            this.ButtonsChanged = false;
+            this.ShouldResetLayout = false;
+            this.Labels.Clear();
+            this.Buttons.Clear();
 
-            this.Bounds = new Rectangle(x, y, width, height);
-            this.OkButton = new ClickableTextureComponent("OK", new Rectangle(x + width - Game1.tileSize * 2, y + height, Game1.tileSize, Game1.tileSize), null, null, Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 46), 1f);
-            this.ClearButton = new ClickableTextureComponent("Cancel", new Rectangle(x + width - Game1.tileSize, y + height, Game1.tileSize, Game1.tileSize), null, null, Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 47), 1f);
-        }
+            // init
+            const int width = KeybindOverlay.BoxWith;
+            const int padding = KeybindOverlay.ContentPadding;
+            const int listIndent = KeybindOverlay.KeybindListIndent;
+            int topOffset = padding;
 
-        /// <summary>Get whether a button is valid for binding.</summary>
-        /// <param name="button">The button to check.</param>
-        private bool IsValidKey(SButton button)
-        {
-            SButton[] blacklist = new[]
+            // add contents with relative positions
+            // (Their real positions will be set once we center the final box.)
             {
-                SButton.ControllerA, // Limitation of how the menu is opened
-                // The rest are because you need to be able to navigate to get to the cancel button...
-                SButton.LeftThumbstickDown,
-                SButton.LeftThumbstickLeft,
-                SButton.LeftThumbstickRight,
-                SButton.LeftThumbstickUp,
-                SButton.RightThumbstickDown,
-                SButton.RightThumbstickLeft,
-                SButton.RightThumbstickRight,
-                SButton.RightThumbstickUp,
-            };
+                // 'Rebinding key: <key>' header
+                string newLine = Environment.NewLine;
+                topOffset += this.AddCenteredLabel(I18n.Config_RebindKey_Title(this.Name) + newLine, 0, 0, topOffset, width).Height;
 
-            return !blacklist.Contains( button );
-        }
+                // edit keybind area
+                bool isEditing = this.KeybindEdit is not null;
+                if (isEditing)
+                {
+                    string text;
+                    if (this.KeybindEdit.Any())
+                        text = this.KeybindEdit.ToString();
+                    else
+                    {
+                        text = this.OnlyAllowSingleButton
+                            ? I18n.Config_RebindKey_SimpleInstructions()
+                            : I18n.Config_RebindKey_ComboInstructions();
+                    }
 
-        /// <summary>Handle the pressed buttons, either by assigning the keybind or cancelling the UI.</summary>
-        /// <param name="buttons">The held or released keys.</param>
-        private void HandleButtons(SButton[] buttons)
-        {
-            // handle escape
-            if (buttons.Any(p => p == SButton.Escape))
-                Game1.playSound("bigDeSelect");
+                    topOffset += this.AddCenteredLabel(text + newLine, 0, 0, topOffset, width).Height;
+                }
+                else if (this.OnlyAllowSingleButton)
+                {
+                    string text = I18n.Config_RebindKey_SimpleInstructions();
 
-            // apply keybind
-            else
-            {
-                Game1.playSound("coin");
-                this.SetValue(buttons);
+                    topOffset += this.AddCenteredLabel(text + newLine, 0, 0, topOffset, width).Height;
+                }
+
+                // keybind list
+                if (!this.OnlyAllowSingleButton && !isEditing)
+                {
+                    // heading
+                    string heading = I18n.Config_RebindKey_KeybindList();
+                    Vector2 headingSize = Game1.dialogueFont.MeasureString(heading);
+                    this.Labels.Add(
+                        new ClickableComponent(new Rectangle(padding, topOffset, (int)headingSize.X, (int)headingSize.Y), string.Empty, heading)
+                    );
+                    topOffset += (int)headingSize.Y;
+
+                    // keybind list
+                    for (int i = 0; i < this.Keybinds.Count; i++)
+                    {
+                        string text = this.Keybinds[i].ToString();
+                        Vector2 size = Game1.dialogueFont.MeasureString(text);
+                        var bounds = new Rectangle(
+                            x: 44 + Game1.tileSize + 10,
+                            y: topOffset,
+                            width: (int)size.X,
+                            height: (int)size.Y
+                        );
+                        if (bounds.Height < Game1.tileSize)
+                            bounds.Y += (Game1.tileSize - bounds.Height) / 2;
+
+                        this.Labels.Add(new ClickableComponent(bounds, string.Empty, text));
+                        this.Buttons.Add(
+                            new ClickableTextureComponent($"{KeybindOverlay.RemoveAction} {i}", new Rectangle(padding + listIndent, topOffset, 44, 44), null, null, Game1.mouseCursors, new Rectangle(338, 494, 11, 11), Game1.pixelZoom)
+                        );
+
+                        topOffset += Math.Max(bounds.Height, 44);
+                    }
+
+                    // add button
+                    var appendButton = new ClickableTextureComponent(KeybindOverlay.AddAction, new Rectangle(padding + listIndent, topOffset, 40, 44), null, null, Game1.mouseCursors, new Rectangle(402, 361, 10, 11), Game1.pixelZoom);
+                    this.Buttons.Add(appendButton);
+                    topOffset += appendButton.bounds.Height;
+                }
             }
+
+            // set content area
+            int height = topOffset + padding;
+            Vector2 pos = Utility.getTopLeftPositionForCenteringOnScreen(width, height);
+            int x = Math.Max(0, (int)pos.X);
+            int y = Math.Max(0, (int)pos.Y);
+            this.Bounds = new Rectangle(x, y, width, height);
+
+            // shift components into content area
+            foreach (ClickableComponent label in this.Labels)
+            {
+                label.bounds.X += x;
+                label.bounds.Y += y;
+            }
+            foreach (ClickableTextureComponent button in this.Buttons)
+            {
+                button.bounds.X += x;
+                button.bounds.Y += y;
+            }
+
+            // add buttons under content box
+            const int mainButtonSize = Game1.tileSize;
+            this.Buttons.AddRange([
+                new ClickableTextureComponent(KeybindOverlay.OkAction, new Rectangle(x + width - mainButtonSize - mainButtonSize, y + height, mainButtonSize, mainButtonSize), null, null, Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 46), 1f),
+                new ClickableTextureComponent(KeybindOverlay.ClearAction, new Rectangle(x + width - mainButtonSize, y + height, mainButtonSize, mainButtonSize), null, null, Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 47), 1f)
+            ]);
         }
 
-        /// <summary>Set the keybind value.</summary>
-        /// <param name="buttons">The buttons to set.</param>
-        private void SetValue(SButton[] buttons)
+        /// <summary>Add a label horizontally centered within the given content area to the <see cref="Labels"/>.</summary>
+        /// <param name="text">The text for which to create a label.</param>
+        /// <param name="contentX">The X pixel position on screen at which the content area begins.</param>
+        /// <param name="contentY">The Y pixel position on screen at which the content area begins.</param>
+        /// <param name="topOffset">The pixel offset from the top of the content area at which to draw the label.</param>
+        /// <param name="contentWidth">The content area's width.</param>
+        /// <returns>Returns the created component's bounds.</returns>
+        private Rectangle AddCenteredLabel(string text, int contentX, int contentY, int topOffset, int contentWidth)
         {
-            switch (this.Option)
+            const int padding = KeybindOverlay.ContentPadding;
+
+            text = Game1.parseText(text, Game1.dialogueFont, contentWidth - padding - padding); // wrap long text
+            Vector2 size = Game1.dialogueFont.MeasureString(text);
+            Rectangle bounds = new(
+                x: (int)(contentX + (contentWidth / 2f) - (size.X / 2f)),
+                y: contentY + topOffset,
+                width: (int)size.X,
+                height: (int)size.Y
+            );
+
+            ClickableComponent label = new ClickableComponent(bounds, "", text);
+            this.Labels.Add(label);
+            return label.bounds;
+        }
+
+        /// <summary>Perform a button action.</summary>
+        /// <param name="action">The button action to perform.</param>
+        private void PerformButtonAction(string action)
+        {
+            switch (action)
             {
-                case SimpleModOption<SButton> opt:
-                    opt.Value = buttons.Any()
-                        ? buttons.Last()
-                        : SButton.None;
-                    this.Label.String = opt.FormatValue();
+                case KeybindOverlay.ClearAction:
+                    Game1.playSound("coin");
+                    this.Keybinds.Clear();
+                    this.Exit();
                     break;
 
-                case SimpleModOption<KeybindList> opt:
-                    opt.Value = new KeybindList(new Keybind(buttons.ToArray()));
-                    this.Label.String = opt.FormatValue();
+                case KeybindOverlay.OkAction:
+                    Game1.playSound("bigDeSelect");
+                    this.Exit();
+                    break;
+
+                case KeybindOverlay.AddAction:
+                    this.StartEditingKeybind(this.Keybinds.Count);
                     break;
 
                 default:
-                    throw new InvalidOperationException($"Unsupported keybind type {typeof(TKeybind).FullName}.");
+                    if (action.StartsWith(KeybindOverlay.RemoveAction))
+                    {
+                        int keyBindIndex = ArgUtility.GetInt(action.Split(' '), 1);
+
+                        this.Keybinds.RemoveAt(keyBindIndex);
+                        this.ShouldResetLayout = true;
+                    }
+                    break;
             }
+        }
+
+        /// <summary>Begin editing a keybind.</summary>
+        /// <param name="index">The index to edit.</param>
+        private void StartEditingKeybind(int index)
+        {
+            this.KeybindEdit = new KeybindEdit(index, this.OnlyAllowSingleButton);
+            this.ShouldResetLayout = true;
+        }
+
+        /// <summary>Finish editing the current keybind.</summary>
+        private void FinishEditingKeybind()
+        {
+            KeybindEdit edit = this.KeybindEdit;
+
+            // save keybind
+            if (edit != null)
+            {
+                Keybind keybind = edit.ToKeybind();
+
+                // add to list
+                if (this.OnlyAllowSingleButton)
+                {
+                    this.Keybinds.Clear();
+                    this.Keybinds.Add(keybind);
+                }
+                else if (edit.Index >= this.Keybinds.Count)
+                    this.Keybinds.Add(keybind);
+                else
+                    this.Keybinds[edit.Index] = keybind;
+
+                // ignore duplicate keybinds
+                if (this.Keybinds.Count > 1)
+                {
+                    HashSet<string> seen = [];
+                    this.Keybinds.RemoveWhere(bind =>
+                        !seen.Add(string.Join(" + ", bind.Buttons.OrderBy(button => button)))
+                    );
+                }
+            }
+
+            // reset
+            this.KeybindEdit = null;
+            this.ShouldResetLayout = true;
+
+            // exit if single button
+            if (this.OnlyAllowSingleButton)
+                this.Exit();
+        }
+
+        /// <summary>Save changes if applicable, and mark the overlay ready to close.</summary>
+        /// <param name="save">Whether to save changes.</param>
+        private void Exit(bool save = true)
+        {
+            if (save)
+                this.OnSaved(this.Keybinds.ToArray());
+
+            this.IsFinished = true;
+            this.KeybindEdit = null;
         }
     }
 }